Add files via upload
Vitiligo Detection using Convolutional Neural Network - This project maily aims to detect the Vitiligo(white) patches in the human skin, when the user uplaods the real-time image of the skin, the processing starts and throws the result whether the user is vitiligo affected or not, apart from it, it also classifies the uploaded image is whether the human skin or any other object also. totally it is a simple user interface that can be accessed by anyone.